Fixed revert not syntax highlighting & memory leak
When you revert a dojo the syntax highlighting is now
re-established.

There was a memory leak in of CodeMirror objects when renaming,
deleting or reverting, this is was removed by not using the
CodeMirror.fromTextArea method and creating the editors using the
DOM element instead.

The problem was that the fromTextArea method added a custom
function to the forms submit and stored the previous function in
itself, when a editor was removed using toTextArea it restores
the submit function that is stored in the editor. If the editors
are not restored in the reverse order that they were added then
this process does not work and memory leaks.