Removed unneeded removeSyntaxHilightEditor method
As the CodeMirror editors are no longer created using the
fromTextArea method, they don't replace the forms submit method
and can be safely removed by just removing the DOM element.
This is already done by other parts of the code so the
removeSyntaxHilightEditor method is no longer needed and has
been removed.