Move traffic-light-count into row with traffic-lights
This is to free up space in the top-left section of the
test and review pages so it can hold more controls.
Plan is for file[new,rename,delete] buttons to move their
on test page, and for fork,revert buttons and diff? checkbox
to move there on review page.

Also drop pie-chart from test and review pages.
Pie-chart now only on dashboard.