Ensure traffic-light count is visible on test page
That is, don't scroll the traffic-light count out of view
when the number of traffic-lights exceeds the max-width of
its container. To do this I dropped the old table-based approach
and used a css flexbox. This solved a secondary traffic-light
ERB problem:
  <% @kata.lights.each do |light| %>
    <%= raw diff_traffic_light(light) %>
  <% end %>
This will generate whitespace only text nodes. These will
affect the layout. Viz, you will get whitespace between the
nodes generated from this when the page is refreshed, but
diff-traffic-lights added by Javascript when you click [test]
will _not_ get whitespace.