Simplify filename :hover CSS handling
This is in preparation for handling hover behaviour explicitly.
The problem is that Alt-J selects a new file but the cursor
(even though it is probably invisibe) is still where it was
(and you cannot move the cursor in JS) so the :hover CSS
means there are two files 'selected'. One for the one that
has actually been selected from the Alt-J hotkey, and one
from where it was before. To fix this I plan to take over
mouseEnter/mouseLeave events and insert/remove a custom
hover-class into the dom node. This will allow the Alt-J
handler to remove this hover-class.
Note that :hover is 'pseudo' and cannot be manipulated in
Javascript.