Fix progress button functionality on dashboard page.
The progress dialog has click handlers on each of its avatar images.
These used css selector .avatar-image but there are also avatar
images on the main rows of the dashboard itself. Furthermore, there
is a heartbeat which auto refreshes and each refresh was adding a new
click handler to all .avatar-image selections, including (incorrectly)
the ones on the progress dialog. So refine the css selector on the
heartbeat.