Turn on lineWrapping for readme.txt, stdout, stderr
This means long lines will wrap. CodeMirror takes care of
the line-numbering.
The readme.txt files in the start-points (eg FizzBuzz)
cannot be edited (yet) to remove all their fixed line-width
newlines because the readme.txt files are what are shown in
the exercises-chooser page when you are setting up a practice
session.