Update how on/off settings cascade...
There is a default in the Javascript, eg fork-button is on.
This can be overriden from the manifest (per kata)
This can be further overriden from an env-var (per server)
However, the env-var from web.env must be set.
For example a web.env entry such as
FORK_BUTTON=
will be ignored.